LayoutTests/imported/w3c/web-platform-tests/html/canvas/element/transformations/2d.transformation.setTransform.nonfinite-expected.txt

2d.transformation.setTransform.nonfinite
setTransform() with Infinity/NaN is ignored
Actual output:
Expected output:


PASS setTransform() with Infinity/NaN is ignored 
